验证码破解/SliderVerification/requirements.txt
numpy
torch==1.9.0
torchvision
matplotlib
tensorflow==2.5.0
tensorboard
terminaltables
pillow
tqdm
